reorganize top bar actions
