+++ feature/topbar/impl/src/main/java/com/arttttt/topbar/impl/components/actions/SettingsTopBarAction.kt
@@ -1,0 +1,10 @@
+package com.arttttt.topbar.impl.components.actions
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Settings
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data object SettingsTopBarAction : TopBarAction {
+
+    override val icon: ImageVector = Icons.Default.Settings
+}
